backend/requirements.txt
asgiref==3.4.1
django==3.2.3
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
requests==2.26.0
django-environ==0.8.1
python-dotenv==0.19.0
six==1.16.0
djoser==2.1.0
webcolors==1.11.1
Pillow==8.4.0
xhtml2pdf==0.2.8
progress==1.6
